# OpenAPI.yaml
openapi: 3.0.0
info:
  title: store
  version: '1.0'
  description: A simple store
  contact:
    name: SecuritEase Dev
    url: 'https://www.securitease.com'
    email: internal@securitease.com
  termsOfService: 'https://www.securitease.com'
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.htm'
servers:
  - url: 'http://localhost:8080'
    description: dev
paths:
  /order:
    get:
      summary: Get Orders
      tags:
        - store
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
      operationId: getOrders
      description: Get all orders
    post:
      summary: Create Order
      tags:
        - store
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
      operationId: createOrder
      description: Create a new order
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
  /order/{id}:
    get:
      summary: Get Order by ID
      tags:
        - store
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Not Found
      operationId: getOrderById
  /customer:
    get:
      summary: Get Customers
      tags:
        - store
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
      operationId: getCustomers
      description: Get all customers
    post:
      summary: Create Customer
      tags:
        - store
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
      operationId: createCustomer
      description: Create a new customer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
  /customer/{name}:
    get:
      summary: Get Customer by Name
      tags:
        - store
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Not Found
      operationId: getCustomerByName
  /products:
    get:
      summary: Get Products
      tags:
        - store
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
      operationId: getProducts
      description: Get all products
    post:
      summary: Create Product
      tags:
        - store
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
      operationId: createProduct
      description: Create a new product
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
  /products/{id}:
    get:
      summary: Get Product by ID
      tags:
        - store
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Not Found
      operationId: getProductById
components:
  schemas:
    Order:
      title: Order
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        customers:
          type: array
          items:
            $ref: '#/components/schemas/OrderCustomer'
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
    Customer:
      title: Customer
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        orders:
          type: array
          items:
            $ref: '#/components/schemas/CustomerOrder'
    Product:
      title: Product
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        orderIds:
          type: array
          items:
            type: integer
            format: int64
    OrderCustomer:
      title: OrderCustomer
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    CustomerOrder:
      title: CustomerOrder
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
tags:
  - name: store
    description: store operations